fix: combine tid2value with pp filter for mds3 tk bug fix
