+++ release.txt
@@ -1,1 +1,2 @@
-
+Fixes:
+- combine tid2value with pp filter for mds3 tk bug fix
